# Eliminate use of method groups in PropertyHelper