Eliminates use of method groups.
Improves performance and reduces allocations.

**Before:**

|             Method |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------- |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|      GetProperties |  36.32 ns | 0.5481 ns | 0.5127 ns |      0.0209 |           - |           - |                88 B |
| ObjectToDictionary | 612.34 ns | 2.0398 ns | 1.8082 ns |      0.1459 |           - |           - |               616 B |

**After:**

|             Method |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------- |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|      GetProperties |  30.36 ns | 0.0877 ns | 0.0820 ns |      0.0057 |           - |           - |                24 B |
| ObjectToDictionary | 605.01 ns | 1.7149 ns | 1.6041 ns |      0.1307 |           - |           - |               552 B |

Benchmark is available [here](https://gist.github.com/drieseng/89ec31cd5a8911ec576852f19cc38803).