@rynowak 2018-11-09
@dotnet-bot test this please