@drieseng 2018-11-13
No need to rush it. I'm kinda waiting for feedback on https://github.com/aspnet/Routing/issues/866. But I've got plenty of other OSS stuff to work on, so no urgency there.